Fixed bug: computation of kruskal-wallis test
If happens that most of the samples are assigned to the same cluster and then no clinical variable is available for the smallest cluster, an error occurs because all samples are part of the same group. If samples having the clinical variable are in the same group, I avoid the computation of the test and assign p-value=1 (i.e. not significant).